working_copy: Break out `TreeConfig` struct and rename `TreeStateSettings`
This is a convenience change that clarifies our model of the tree state and its
initialization. This creates a useful separation between the public settings
type (`LocalWcSettings`), which is used elsewhere in the CLI, and the private
internal configuration of the tree state `TreeStateConfig` which needs more
initialization besides just accessing settings.